refactor: extract rate limiter configuration into dedicated module #531
